% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conc.R
\name{calc_conc}
\alias{calc_conc}
\title{Calculate Concentrations for Sample Set}
\usage{
calc_conc(samples, curve, dilution = 1)
}
\arguments{
\item{samples}{Takes in a dataframe with processed sample information and converts the absorbance values to concentrations.}

\item{curve}{Takes in a vector object that contains the calibration curve information from calibrate_set().}

\item{dilution}{Takes in a numeric value that is used to determine and dilution factors to include.}
}
\description{
Calculates concentrations of samples based on the calibration curve created from calibrate_set()
}
