% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_set.R
\name{calibrate_set}
\alias{calibrate_set}
\title{Get Calibration Curve from Standards}
\usage{
calibrate_set(stds)
}
\arguments{
\item{stds}{Takes in a dataframe that corresponds to the standards data among the data sets.}
}
\description{
Creates a linear model from standard data
}
