% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_set.R
\name{check_set}
\alias{check_set}
\title{Check that the set is valid}
\usage{
check_set(df)
}
\arguments{
\item{df}{Takes in a dataframe object that corresponds to the sample or standard files}
}
\description{
Checks if the filename column is explicitly written and ensures the format is correct
}
