% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_abs.R
\name{find_abs}
\alias{find_abs}
\title{Find Absorbance Value}
\usage{
find_abs(txtfile)
}
\arguments{
\item{txtfile}{Takes in a character value that corresponds to a textfile with spectral data}

\item{wl}{Takes in a numeric value that corresponds to the wavelength you would like to extract absorbance values from. The default value is 562 for FerroZine assays}
}
\description{
Retrieves the absorbance value for FerroZine measurements
}
