% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spec.R
\name{get_spec}
\alias{get_spec}
\title{Retrieve Spectral Values}
\usage{
get_spec(txtfile)
}
\arguments{
\item{txtfile}{Takes in a character that corresponds to a text file that contains spectral data. This is designed to parse the structure from the collaborator's sepctrophotometer.}
}
\value{
A dataframe with wavelengths and absorbance
}
\description{
get_spec() parses a text file to extract only the absorption and wavelength values
}
